backend/requirements.txt
fastapi==0.115.6
gunicorn
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.4
